requirements.txt
amqp==2.2.2
babel==2.5.3
backcall==0.1.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
django-authtools==1.6.0
django-bootstrap4==0.0.6
django-extensions==2.0.7
django-filter==1.1.0
django-taggit==0.22.2
django==2.0.4
idna==2.6
inflect==0.2.5
ipython-genutils==0.2.0
ipython==6.3.1
isodate==0.6.0
jedi==0.12.0
kombu==4.1.0
mwoauth==0.3.2
numpy==1.14.2
oauthlib==2.0.7
pandas==0.22.0
parso==0.2.0
pexpect==4.5.0; sys_platform != 'win32'
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pygments==2.2.0
pyjwt==1.6.1
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
rdflib==4.2.2
requests-oauthlib==0.8.0
requests==2.18.4
silly==1.1.0
simplegeneric==0.8.1
six==1.11.0
sparqlwrapper==1.8.1
tqdm==4.23.0
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
wikidata==0.6.1
wikipedia==1.4.0
